extract bothel roll data
