+++ src/MiniEstablishments/Brothel/brothelData.ts
@@ -5,12 +5,6 @@ import { createRelationship } from '../../NPCGeneration/Relationships/createRela
 import { createParentage } from '../../NPCGeneration/Relationships/createFamilyMembers'
 
 interface BrothelData {
-  rollData: {
-    wealth: RollData
-    size: RollData
-    cleanliness: RollData
-    bedCleanliness: RollData
-  }
   name: string[]
   /**
    * @example "Apparently, it specializes in ____"
@@ -41,12 +35,6 @@ interface BrothelData {
   harlot: HarlotData
 }
 
-interface RollData {
-  description: string
-  preceding: string
-  rolls: [number, string][]
-}
-
 interface HarlotData {
   create(town: Town, brothel: Building, base: Partial<NPC>): string
   type: Record<string, HarlotTypeData>
@@ -69,69 +57,6 @@ interface HarlotTypeData {
 }
 
 export const brothelData: BrothelData = {
-  rollData: {
-    wealth: {
-      description: 'How successful is the brothel?',
-      preceding: 'Brothel Wealth:',
-      rolls: [
-        [95, 'kingly'],
-        [80, 'aristocratic'],
-        [70, 'wealthy'],
-        [60, 'comfortable'],
-        [50, 'modest'],
-        [25, 'poor'],
-        [15, 'squalid'],
-        [0, 'destitute']
-      ]
-    },
-    size: {
-      description: 'How large is the brothel?',
-      preceding: 'Brothel Size:',
-      rolls: [
-        [95, 'cavernous'],
-        [80, 'huge'],
-        [70, 'quite large'],
-        [60, 'large'],
-        [50, 'spacious'],
-        [40, 'average sized'],
-        [30, 'somewhat cramped'],
-        [20, 'small'],
-        [10, 'tiny'],
-        [0, 'extremely cramped']
-      ]
-    },
-    cleanliness: {
-      description: 'How clean is the brothel?',
-      preceding: 'Brothel Cleanliness:',
-      rolls: [
-        [80, 'fastidious'],
-        [70, 'very tidy'],
-        [60, 'tidy'],
-        [50, 'reasonably tidy'],
-        [40, 'somewhat messy'],
-        [30, 'rather messy'],
-        [20, 'very messy'],
-        [10, 'extremely messy'],
-        [0, 'dangerously messy']
-      ]
-    },
-    bedCleanliness: {
-      description: 'How clean are the brothel beds?',
-      preceding: 'Brothel Bed Cleanliness:',
-      rolls: [
-        [80, 'perfectly prepared, with fresh sheets and a lemon scent in the air of the room'],
-        [75, 'recently prepared and well cleaned'],
-        [70, 'freshly cleaned and neat'],
-        [60, 'tidy and neat'],
-        [50, 'reasonably clean'],
-        [40, 'somewhat tidy'],
-        [30, 'disgusting'],
-        [20, 'teeming with rats'],
-        [10, 'rather filthy'],
-        [0, 'festering with bugs']
-      ]
-    }
-  },
   name: [
     'Daisies', 'The Don', 'Blinkers', "The Prude's Suspenders", 'Gold’s Dust', 'The Velvet Fang', 'The Whisper Home', 'The Hook', 'Over the River', 'Slooshes', 'The Rapid Rascal', 'The Stoat',
     'The Fidgety Ferret', 'Long Shaft', 'The Guards', 'Chastity’s', 'The Mask', 'The Red Dress', 'Jewels', 'Silken Flute', 'The Nag’s Head', 'Drummers', 'Sailors Choice', 'Blacksmiths Envy',
